feat: added controllers folder and seperated the from route
