+++ routes/customers.js
@@ -1,56 +1,14 @@
-import { Customer, validateCustomer } from '../models/customer.js';
+import * as customersController from '../controllers/customers.js';
 import * as express from 'express';
 const router = express.Router();
 
-router.get('/', async (req, res) => {
-  const customers = await Customer.find().sort('name').lean().select('-__v');
-  res.send(customers);
-});
+router.route('/')
+  .get(customersController.getCustomers)
+  .post(customersController.storeCustomer);
 
-router.post('/', async (req, res) => {
-  const { error } = validateCustomer(req.body); 
-  if (error) return res.status(400).send(error.details[0].message);
-
-  let customer = new Customer({ 
-    name: req.body.name,
-    isGold: req.body.isGold,
-    phone: req.body.phone
-  });
-  customer = await customer.save();
-  
-  res.send(customer);
-});
-
-router.put('/:id', async (req, res) => {
-  const { error } = validateCustomer(req.body); 
-  if (error) return res.status(400).send(error.details[0].message);
-
-  const customer = await Customer.findByIdAndUpdate(req.params.id,
-    { 
-      name: req.body.name,
-      isGold: req.body.isGold,
-      phone: req.body.phone
-    }, { new: true }).lean().select('-__v');
-
-  if (!customer) return res.status(404).send('The customer with the given ID was not found.');
-  
-  res.send(customer);
-});
-
-router.delete('/:id', async (req, res) => {
-  const customer = await Customer.findByIdAndRemove(req.params.id).lean().select('-__v');
-
-  if (!customer) return res.status(404).send('The customer with the given ID was not found.');
-
-  res.send(customer);
-});
-
-router.get('/:id', async (req, res) => {
-  const customer = await Customer.findById(req.params.id).lean().select('-__v -date');
-
-  if (!customer) return res.status(404).send('The customer with the given ID was not found.');
-
-  res.send(customer);
-});
+router.route('/:id')
+  .put(customersController.updateCustomer)
+  .delete(customersController.deleteCustomer)
+  .get(customersController.getCustomer);
 
 export default router;
