+++ routes/movies.js
@@ -1,71 +1,14 @@
-import { Movie, validateMovie } from '../models/movie.js';
-import { Genre } from '../models/genre.js';
+import * as moviesController from '../controllers/movies.js';
 import * as express from 'express';
 const router = express.Router();
 
-router.get('/', async (req, res) => {
-  const movies = await Movie.find().sort('name').lean().select('-__v');
-  res.send(movies);
-});
+router.route('/')
+  .get(moviesController.getMovies)
+  .post(moviesController.storeMovie);
 
-router.post('/', async (req, res) => {
-  const { error } = validateMovie(req.body); 
-  if (error) return res.status(400).send(error.details[0].message);
-
-  const genre = await Genre.findById(req.body.genreId).lean();
-  if (!genre) return res.status(400).send('Invalid genre.');
-
-  const movie = new Movie({
-    title: req.body.title,
-    genre: {
-      _id: genre._id,
-      name: genre.name
-    },
-    numberInStock: req.body.numberInStock,
-    dailyRentalRate: req.body.dailyRentalRate
-  });
-  await movie.save();
-  
-  res.send(movie);
-});
-
-router.put('/:id', async (req, res) => {
-  const { error } = validateMovie(req.body); 
-  if (error) return res.status(400).send(error.details[0].message);
-
-  const genre = await Genre.findById(req.body.genreId).lean();
-  if (!genre) return res.status(400).send('Invalid genre.');
-
-  const movie = await Movie.findByIdAndUpdate(req.params.id,
-    { 
-      title: req.body.title,
-      genre: {
-        _id: genre._id,
-        name: genre.name
-      },
-      numberInStock: req.body.numberInStock,
-      dailyRentalRate: req.body.dailyRentalRate
-    }, { new: true }).lean();
-
-  if (!movie) return res.status(404).send('The movie with the given ID was not found.');
-  
-  res.send(movie);
-});
-
-router.delete('/:id', async (req, res) => {
-  const movie = await Movie.findByIdAndRemove(req.params.id).lean().select('-__v');
-
-  if (!movie) return res.status(404).send('The movie with the given ID was not found.');
-
-  res.send(movie);
-});
-
-router.get('/:id', async (req, res) => {
-  const movie = await Movie.findById(req.params.id).lean().select('-__v');
-
-  if (!movie) return res.status(404).send('The movie with the given ID was not found.');
-
-  res.send(movie);
-});
+router.route('/:id')
+  .put(moviesController.updateMovie)
+  .delete(moviesController.deleteMovie)
+  .get(moviesController.getMovie);
 
 export default router;
