+++ routes/users.js
@@ -1,29 +1,9 @@
+import * as usersController from '../controllers/users.js';
 import auth from '../middleware/auth.js';
-import * as bcrypt from 'bcrypt';
-import _ from 'lodash';
-import { User, validateUser } from '../models/user.js';
 import * as express from 'express';
 const router = express.Router();
 
-router.get('/me', auth, async (req, res) => {
-  const user = await User.findById(req.user._id).select('-password').lean();
-  res.send(user);
-});
-
-router.post('/', async (req, res) => {
-  const { error } = validateUser(req.body); 
-  if (error) return res.status(400).send(error.details[0].message);
-
-  let user = await User.findOne({ email: req.body.email });
-  if (user) return res.status(400).send('User already registered.');
-
-  user = new User(_.pick(req.body, ['name', 'email', 'password']));
-  const salt = await bcrypt.genSalt(10);
-  user.password = await bcrypt.hash(user.password, salt);
-  await user.save();
-
-  const token = user.generateAuthToken();
-  res.header('x-auth-token', token).send(_.pick(user, ['_id', 'name', 'email']));
-});
+router.get('/me', auth, usersController.getCurrentUser);
+router.post('/', usersController.createUser);
 
 export default router;
